// src/components/importarOlimpista/types/olympista.ts
export type Olympian = {
  id: string;
  TIPO_PART: string;
  AREA_COD: string;
  AREA_NOM: string;
  NIVEL_COD: string;
  NIVEL_NOM: string;
  OLI_TDOC: string;
  OLI_NRODOC: string;
  OLI_NOMBRE: string;
  OLI_AP_PAT: string;
  OLI_AP_MAT: string;
  OLI_UNID_EDU: string;
  OLI_DEPTO: string;
  OLI_GRADO: string;
  OLI_F_NAC: string;
  OLI_SEXO: string;
  OLI_CORREO: string;
  TUTOR_TDOC: string;
  TUTOR_NRODOC: string;
  TUTOR_NOMBRE: string;
  TUTOR_AP_PAT: string;
  TUTOR_AP_MAT: string;
  TUTOR_TEL: string;
  TUTOR_CORREO: string;
  TUTOR_UNID_EDU: string;
  TUTOR_PROF: string;
  EQUIPO_NOMBRE: string;
  ROL_EQUIPO: string;
};